ZCS-8785:Fix for issue with data type range values for Mysql

+++ client/src/java/com/zimbra/client/event/ZModifyFolderEvent.java
@@ -158,8 +158,8 @@ public class ZModifyFolderEvent implements ZModifyItemEvent, ToZJSONObject {
      * @return new content sequence or defaultValue if unchanged
      * @throws ServiceException
      */
-    public int getContentSequence(int defaultValue) throws ServiceException {
-        return (int) mFolderEl.getAttributeLong(MailConstants.A_REVISION, defaultValue);
+    public long getContentSequence(long defaultValue) throws ServiceException {
+        return mFolderEl.getAttributeLong(MailConstants.A_REVISION, defaultValue);
     }
 
     /**
